+++ CHANGELOG.md
@@ -2,5 +2,9 @@
 11 december 2022:
 - start of devlog
 - changed build process a bit.
-- added GUI (powered by nuklear (it's so bad :( ))
+- added GUI (powered by nuklear)
+
+17 decenmber 2022:
+- added New Session Manager (any OSC) support
+- **breaked compability with old clients** : `ju_ctx_init()` arguments changed!